requirements.txt
PyOpenGL
PyGLUT
PyOpenGL-accelerate
glfw
numpy